requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
Django==5.2
easydict==1.13
flatbuffers==25.2.10
fonttools==4.57.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
onnx==1.17.0
onnxruntime==1.21.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
prettytable==3.16.0
protobuf==6.30.2
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
sqlparse==0.5.3
stringzilla==3.12.5
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.30
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
